refactor erddapy.py and remove url_builder.py
