+++ erddapy/erddapy.py
@@ -5,17 +5,20 @@ Pythonic way to access ERDDAP data
 
 """
 
+import copy
 import functools
+from urllib.parse import quote_plus
 
 import pandas as pd
 
-from erddapy.url_builder import (
-    categorize_url,
-    download_url,
-    info_url,
-    search_url,
+from erddapy.utilities import (
+    _check_url_response,
+    _tempnc,
+    parse_dates,
+    quote_string_constraints,
+    servers,
+    urlopen,
 )
-from erddapy.utilities import _check_url_response, _tempnc, servers, urlopen
 
 
 class ERDDAP(object):
@@ -106,17 +109,109 @@ class ERDDAP(object):
         page=1,
         **kwargs,
     ):
+
+        """The search URL for the `server` endpoint provided.
+
+        Args:
+            search_for (str): "Google-like" search of the datasets' metadata.
+
+                - Type the words you want to search for, with spaces between the words.
+                    ERDDAP will search for the words separately, not as a phrase.
+                - To search for a phrase, put double quotes around the phrase
+                    (for example, `"wind speed"`).
+                - To exclude datasets with a specific word, use `-excludedWord`.
+                - To exclude datasets with a specific phrase, use `-"excluded phrase"`
+                - Searches are not case-sensitive.
+                - To find just grid or just table datasets, include `protocol=griddap`
+                    or `protocol=tabledap` in your search.
+                - You can search for any part of a word. For example,
+                    searching for `spee` will find datasets with `speed` and datasets with
+                    `WindSpeed`
+                - The last word in a phrase may be a partial word. For example,
+                    to find datasets from a specific website (usually the start of the datasetID),
+                    include (for example) `"datasetID=erd"` in your search.
+
+            response (str): default is HTML.
+            items_per_page (int): how many items per page in the return,
+                default is 1000.
+            page (int): which page to display, default is the first page (1).
+            kwargs (dict): extra search constraints based on metadata and/or coordinates ke/value.
+                metadata: `cdm_data_type`, `institution`, `ioos_category`,
+                `keywords`, `long_name`, `standard_name`, and `variableName`.
+                coordinates: `minLon`, `maxLon`, `minLat`, `maxLat`, `minTime`, and `maxTime`.
+
+        Returns:
+            url (str): the search URL.
+
+        """
+        base = (
+            "{server}/search/advanced.{response}"
+            "?page={page}"
+            "&itemsPerPage={itemsPerPage}"
+            "&protocol={protocol}"
+            "&cdm_data_type={cdm_data_type}"
+            "&institution={institution}"
+            "&ioos_category={ioos_category}"
+            "&keywords={keywords}"
+            "&long_name={long_name}"
+            "&standard_name={standard_name}"
+            "&variableName={variableName}"
+            "&minLon={minLon}"
+            "&maxLon={maxLon}"
+            "&minLat={minLat}"
+            "&maxLat={maxLat}"
+            "&minTime={minTime}"
+            "&maxTime={maxTime}"
+        )
+        if search_for:
+            search_for = quote_plus(search_for)
+            base += "&searchFor={searchFor}"
+
+        # Convert dates from datetime to `seconds since 1970-01-01T00:00:00Z`.
+        min_time = kwargs.pop("min_time", None)
+        max_time = kwargs.pop("max_time", None)
+        if min_time:
+            kwargs.update({"min_time": parse_dates(min_time)})
+        if max_time:
+            kwargs.update({"max_time": parse_dates(max_time)})
+
+        default = "(ANY)"
         response = response if response else self.response
-        return search_url(
+        url = base.format(
             server=self.server,
             response=response,
-            search_for=search_for,
-            items_per_page=items_per_page,
             page=page,
-            **kwargs,
+            itemsPerPage=items_per_page,
+            protocol=kwargs.get("protocol", default),
+            cdm_data_type=kwargs.get("cdm_data_type", default),
+            institution=kwargs.get("institution", default),
+            ioos_category=kwargs.get("ioos_category", default),
+            keywords=kwargs.get("keywords", default),
+            long_name=kwargs.get("long_name", default),
+            standard_name=kwargs.get("standard_name", default),
+            variableName=kwargs.get("variableName", default),
+            minLon=kwargs.get("min_lon", default),
+            maxLon=kwargs.get("max_lon", default),
+            minLat=kwargs.get("min_lat", default),
+            maxLat=kwargs.get("max_lat", default),
+            minTime=kwargs.get("min_time", default),
+            maxTime=kwargs.get("max_time", default),
+            searchFor=search_for,
         )
 
+        return _check_url_response(url)
+
     def get_info_url(self, dataset_id=None, response=None):
+        """The info URL for the `server` endpoint.
+
+        Args:
+            dataset_id (str): a dataset unique id.
+            response (str): default is HTML.
+
+        Returns:
+            url (str): the info URL for the `response` chosen.
+
+        """
         dataset_id = dataset_id if dataset_id else self.dataset_id
         response = response if response else self.response
 
@@ -125,15 +220,27 @@ class ERDDAP(object):
                 f"You must specify a valid dataset_id, got {dataset_id}"
             )
 
-        return info_url(
-            server=self.server, dataset_id=dataset_id, response=response
-        )
+        url = f"{self.server}/info/{dataset_id}/index.{response}"
+        return _check_url_response(url)
 
     def get_categorize_url(self, categorize_by, value=None, response=None):
+        """The categorize URL for the `server` endpoint.
+
+        Args:
+            categorize_by (str): a valid attribute, e.g.: ioos_category or standard_name.
+            value (str): an attribute value.
+            response (str): default is HTML.
+
+        Returns:
+            url (str): the categorized URL for the `response` chosen.
+
+        """
         response = response if response else self.response
-        return categorize_url(
-            self.server, categorize_by, value=value, response=response
-        )
+        if value:
+            url = f"{self.server}/categorize/{categorize_by}/{value}/index.{response}"
+        else:
+            url = f"{self.server}/categorize/{categorize_by}/index.{response}"
+        return _check_url_response(url)
 
     def get_download_url(
         self,
@@ -143,6 +250,26 @@ class ERDDAP(object):
         response=None,
         constraints=None,
     ):
+        """The download URL for the `server` endpoint.
+
+        Args:
+            dataset_id (str): a dataset unique id.
+            protocol (str): tabledap or griddap.
+            variables (list/tuple): a list of the variables to download.
+            response (str): default is HTML.
+            constraints (dict): download constraints, default None (opendap-like url)
+                example: constraints = {'latitude<=': 41.0,
+                                        'latitude>=': 38.0,
+                                        'longitude<=': -69.0,
+                                        'longitude>=': -72.0,
+                                        'time<=': '2017-02-10T00:00:00+00:00',
+                                        'time>=': '2016-07-10T00:00:00+00:00',
+                                        }
+
+        Returns:
+            url (str): the download URL for the `response` chosen.
+
+        """
         dataset_id = dataset_id if dataset_id else self.dataset_id
         protocol = protocol if protocol else self.protocol
         variables = variables if variables else self.variables
@@ -159,14 +286,29 @@ class ERDDAP(object):
                 f"Please specify a valid `protocol`, got {protocol}"
             )
 
-        return download_url(
-            server=self.server,
-            dataset_id=dataset_id,
-            protocol=protocol,
-            variables=variables,
-            response=response,
-            constraints=constraints,
-        )
+        # This is an unconstrained OPeNDAP response b/c
+        # the integer based constrained version is just not worth supporting ;-p
+        if response == "opendap":
+            return f"{self.server}/{protocol}/{dataset_id}"
+        else:
+            url = f"{self.server}/{protocol}/{dataset_id}.{response}?"
+
+        if variables:
+            variables = ",".join(variables)
+            url += f"{variables}"
+
+        if constraints:
+            _constraints = copy.copy(constraints)
+            for k, v in _constraints.items():
+                if k.startswith("time"):
+                    _constraints.update({k: parse_dates(v)})
+            _constraints = quote_string_constraints(_constraints)
+            _constraints = "".join(
+                [f"&{k}{v}" for k, v in _constraints.items()]
+            )
+
+            url += f"{_constraints}"
+        return _check_url_response(url)
 
     def to_pandas(self, **kw):
         """Save a data request to a pandas.DataFrame.
@@ -220,7 +362,9 @@ class ERDDAP(object):
                 f"You must specify a valid dataset_id, got {dataset_id}"
             )
 
-        url = info_url(self.server, dataset_id=dataset_id, response="csv")
+        url = self.get_info_url(
+            self.server, dataset_id=dataset_id, response="csv"
+        )
 
         variables = {}
         _df = pd.read_csv(
